requirements.txt
Flask==2.0.1
psycopg2-binary==2.9.1
requests==2.26.0
python-dotenv==0.19.0
beautifulsoup4==4.9.3
pyperclip==1.8.2